pyproject.toml
[project]
name = "legix"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "sqlalchemy[asyncio]",
    "aiosqlite",
    "anthropic",
    "pydantic-settings",
    "langchain",
    "langchain-anthropic",
    "langgraph",
]

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = [".."]
include = ["legix*"]
